+++ lib/graphql/client.ex
@@ -12,6 +12,11 @@ defmodule BridgeEx.Graphql.Client do
           | {:error, {:http_error, String.t()}}
           | {:error, list()}
 
+  @http_options timeout: 1_000, recv_timeout: 16_000
+  @http_headers %{
+    "Content-type" => "application/json"
+  }
+
   @doc """
   Calls a GraphQL endpoint
 
@@ -19,9 +24,14 @@ defmodule BridgeEx.Graphql.Client do
 
     * `url`: URL of the endpoint.
     * `query`: Graphql query or mutation.
-    * `variables`: dariables for Graphql query or mutation.
-    * `http_options`: HTTPoison options.
-    * `http_headers`: HTTPoison headers.
+    * `variables`: variables for Graphql query or mutation.
+    * `opts`: various options.
+
+  ## Options
+
+    * `options`: extra HTTP options to be passed to Telepoison.
+    * `headers`: extra HTTP headers.
+    * `encode_variables`: whether to JSON encode variables or not.
     * `retry_options`: configures retry attempts. Takes the form of `[max_retries: 1, timing: :exponential]`
     * `log_options`: configures logging on errors. Takes the form of `[log_query_on_error: false, log_response_on_error: false]`.
   """
@@ -29,21 +39,40 @@ defmodule BridgeEx.Graphql.Client do
   @spec call(
           url :: String.t(),
           query :: String.t(),
-          variables :: map() | String.t(),
-          http_options :: Keyword.t(),
-          http_headers :: map(),
-          retry_options :: Keyword.t(),
-          log_options :: Keyword.t()
+          variables :: map(),
+          opts :: Keyword.t()
         ) :: bridge_response()
   def call(
         url,
         query,
         variables,
-        http_options,
-        http_headers,
-        retry_options,
-        log_options
+        opts
       ) do
+    encode_variables = Keyword.get(opts, :encode_variables, false)
+    http_options = Keyword.merge(@http_options, Keyword.get(opts, :options, []))
+    http_headers = Map.merge(@http_headers, Keyword.get(opts, :headers, %{}))
+    log_options = Keyword.merge(log_options(), Keyword.get(opts, :log_options))
+
+    retry_options =
+      opts
+      |> Keyword.get(:retry_options, [])
+      |> then(
+        &Keyword.merge(
+          [
+            delay: 100,
+            max_retries: 0,
+            policy: fn _ -> true end,
+            timing: :exponential
+          ],
+          &1
+        )
+      )
+
+    variables =
+      if encode_variables,
+        do: Jason.encode!(variables),
+        else: variables
+
     %{query: String.trim(query), variables: variables}
     |> Jason.encode()
     |> Noether.Either.bind(
@@ -71,4 +100,11 @@ defmodule BridgeEx.Graphql.Client do
 
   def format_response(response) when is_map(response), do: Utils.normalize_inner_fields(response)
   def format_response(response), do: response
+
+  defp log_options do
+    Application.get_env(:bridge_ex, :log_options,
+      log_query_on_error: false,
+      log_response_on_error: false
+    )
+  end
 end