+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Next]
 
+## [2.0.0] - 2022-05-19
+
+### Changed
+
+- **Breaking**: `BridgeEx.Graphql.Client.call` now accepts an `opts :: Keyword.t()` parameter instead of specific options
+
+### Fixed
+
+- `BridgeEx.Graphql.Client.call` accepts only `variables :: map()` once again
+- `BridgeEx.Graphql.Client.call` now performs `variables` encoding internally
+
 ## [1.2.0] - 2022-05-16
 
 ### Added
@@ -99,7 +110,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of `bridge_ex`
 
-[Next]: https://github.com/primait/bridge_ex/compare/1.1.0...HEAD
+[Next]: https://github.com/primait/bridge_ex/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/primait/bridge_ex/compare/1.2.0...2.0.0
+[1.2.0]: https://github.com/primait/bridge_ex/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/primait/bridge_ex/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/primait/bridge_ex/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/primait/bridge_ex/compare/0.4.1...1.0.0